feature(pdf) Add browser version of PdfTransformer
