+++ packages/markdown-pdf/src/PdfTransformer.js
@@ -14,61 +14,13 @@
 
 'use strict';
 
-const Stream = require('stream');
-
-const ToPdfMake = require('./ToPdfMake');
 const ToCiceroMark = require('./ToCiceroMark');
-
-const PdfPrinter = require('pdfmake');
-const { defaultFonts } = require('./pdfmakeutil');
+const PdfTransformerBase = require('./PdfTransformerBase');
 
 /**
  * Converts a PDF to CiceroMark DOM
  */
-class PdfTransformer {
-    /**
-     * Generates a buffer for a pdf document
-     * @param {object} input - the pdfmake DOM
-     * @param {*} t - the transform
-     * @return {*} a promise to a buffer
-     */
-    static toBuffer(input, t) {
-        const outputStream = new Stream.Writable();
-
-        let arrayBuffer = new ArrayBuffer(8);
-        let memStore = Buffer.from(arrayBuffer);
-        outputStream._write = (chunk, encoding, next) => {
-            let buffer = (Buffer.isBuffer(chunk))
-                ? chunk  // already is Buffer use it
-                : new Buffer(chunk, encoding);  // string, convert
-
-            // concat to the buffer already there
-            memStore = Buffer.concat([memStore, buffer]);
-            next();
-        };
-
-        return new Promise((resolve) => {
-            outputStream.on('finish', () => {
-                resolve(memStore);
-            });
-            // Call the transform, passing the stream
-            return t(input, outputStream);
-        });
-    }
-
-    /**
-     * Generates a base64 string for a pdf document
-     * @param {object} input - the pdfmake DOM
-     * @param {*} t - the transform
-     * @return {string} a base64 encoded string for the pdf
-     */
-    static toBase64(input, t) {
-        return PdfTransformer.toBuffer(input, t).then((buffer) => {
-            const bytes = Buffer.from(buffer);
-            return bytes.toString('base64');
-        });
-    }
-
+class PdfTransformer extends PdfTransformerBase {
     /**
      * Converts an pdf buffer to a CiceroMark DOM
      * @param {Buffer} input - pdf buffer
@@ -87,74 +39,13 @@ class PdfTransformer {
     }
 
     /**
-     * Converts a CiceroMark DOM to a pdfmake DOM
-     * @param {*} input - CiceroMark DOM (JSON)
-     * @param {*} options - the PDF generation options
-     * @param {boolean} [options.saveCiceroMark] - whether to save source CiceroMark as a custom property (defaults to true)
-     * @param {array} [options.templates] - an array of buffers to be saved into the PDF as custom base64 encoded properties (defaults to null)
-     * @return {*} the pdfmake DOM
-     */
-    static async ciceroMarkToPdfMake(input, options = { saveCiceroMark: true }) {
-        // The JSON document in pdfmake format
-        const dd = await ToPdfMake.CiceroMarkToPdfMake(input, options);
-        return dd;
-    }
-
-    /**
-     * Converts a TemplateMark DOM to a pdfmake DOM
-     * @param {*} input - TemplateMark DOM (JSON)
-     * @param {*} options - the PDF generation options
-     * @param {boolean} [options.saveCiceroMark] - whether to save source CiceroMark as a custom property (defaults to true)
-     * @param {array} [options.templates] - an array of buffers to be saved into the PDF as custom base64 encoded properties (defaults to null)
-     * @return {*} the pdfmake DOM
-     */
-    static async templateMarkToPdfMake(input, options = { saveTemplateMark: true }) {
-        // The JSON document in pdfmake format
-        const dd = await ToPdfMake.TemplateMarkToPdfMake(input, options);
-        return dd;
-    }
-
-    /**
-     * Converts a pdfmake DOM to a PDF Buffer
-     * @param {*} progressCallback - a callback function used during pdf emit
-     * @return {*} a function from input and stream, adding the pdf to the stream
-     */
-    static pdfMakeToPdfStreamWithCallback(progressCallback) {
-        return (input, outputStream) => {
-            // Default fonts are better defined at rendering time
-            input.defaultStyle = {
-                fontSize: 12,
-                font: 'LiberationSerif',
-                lineHeight: 1.5
-            };
-
-            // The Pdf printer
-            const printer = new PdfPrinter(defaultFonts);
-
-            // Printing to stream
-            const pdfDoc = printer.createPdfKitDocument(input, { progressCallback });
-            pdfDoc.pipe(outputStream);
-            pdfDoc.end();
-        };
-    }
-
-    /**
-     * Converts a pdfmake DOM to a PDF Buffer
-     * @param {*} input - pdfmake DOM (JSON)
-     * @param {*} outputStream - the output stream
-     */
-    static async pdfMakeToPdfStream(input, outputStream) {
-        return PdfTransformer.pdfMakeToPdfStreamWithCallback()(input, outputStream);
-    }
-
-    /**
      * Converts a pdfmake DOM to a PDF Buffer
      * @param {*} input - pdfmake DOM (JSON)
      * @param {*} progressCallback - a callback function used during pdf emit
      * @return {*} a pdf buffer
      */
     static async pdfMakeToPdfBuffer(input, progressCallback) {
-        return PdfTransformer.toBuffer(input, PdfTransformer.pdfMakeToPdfStreamWithCallback(progressCallback));
+        return PdfTransformerBase.toBuffer(input, PdfTransformerBase.pdfMakeToPdfStreamWithCallback(progressCallback));
     }
 }
